+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### [2.0.0-rc.62] - 2025-10-04
+
+### Added
+
+- bootstrap icons
+- interactive disclosure! (not in backend yet)
+- full rebuild on actionLisp to allow declare and identifyAs actions
+- button payloads on Menu too
+
 ### [2.0.0-rc.61] - 2025-10-02
 
 ### Fixed